+++ packages/workspace-core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.2](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-core@0.1.1...@equinor/workspace-core@0.1.2) (2023-02-06)
+
+**Note:** Version bump only for package @equinor/workspace-core
+
 ## [0.1.1](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-core@0.1.0...@equinor/workspace-core@0.1.1) (2022-12-12)
 
 **Note:** Version bump only for package @equinor/workspace-core