+++ packages/workspace-fusion/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.1](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.17.0...@equinor/workspace-fusion@0.17.1) (2023-02-08)
+
+**Note:** Version bump only for package @equinor/workspace-fusion
+
+## [0.17.0](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.16.6...@equinor/workspace-fusion@0.17.0) (2023-02-07)
+
+### Features
+
+-   ✨workspace-dynamic ([#317](https://github.com/equinor/fusion-workspace/issues/317)) ([f53e16d](https://github.com/equinor/fusion-workspace/commit/f53e16dae70fb7b317ba875da93733e6ea1f69d3)), closes [#2](https://github.com/equinor/fusion-workspace/issues/2)
+
+## [0.16.6](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.16.5...@equinor/workspace-fusion@0.16.6) (2023-02-06)
+
+**Note:** Version bump only for package @equinor/workspace-fusion
+
+## [0.16.5](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.16.4...@equinor/workspace-fusion@0.16.5) (2023-02-06)
+
+**Note:** Version bump only for package @equinor/workspace-fusion
+
 ## [0.16.4](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-fusion@0.16.3...@equinor/workspace-fusion@0.16.4) (2023-02-03)
 
 **Note:** Version bump only for package @equinor/workspace-fusion